REXFLOW-110- (productionize step 1)

+++ deploy/deployer.py
@@ -3,6 +3,7 @@ import logging
 import kubernetes.client
 import kubernetes.client.rest
 import os
+import subprocess
 
 from . import specs
 
@@ -50,7 +51,10 @@ class Deployer:
         self.delete_namespaced_custom_object = wrap_api_call(
             self.custom_api.delete_namespaced_custom_object)
 
-    def create(self, _):
+    def create(self, namespace):
+        print("The deploy module is used for dev deployments. As such, we are now "
+              "setting the kube context to docker-desktop.", flush=True)
+        subprocess.check_output("kubectl config use-context docker-desktop".split())
         self.create_namespace(specs.rexflow_namespace_spec)
         # ETCD
         self.create_namespaced_service_account(
@@ -67,7 +71,9 @@ class Deployer:
         self.create_namespaced_service(
             'rexflow', specs.flowd_service_specs['rexflow'])
         self.create_namespaced_deployment(
-            'rexflow', specs.mk_flowd_deployment_spec('rexflow-etcd.rexflow'))
+            'rexflow', specs.mk_flowd_deployment_spec('rexflow-etcd.rexflow',
+            namespace.kafka
+        ))
         self.create_namespaced_role_binding(
             'default', specs.flowd_edit_default_spec)
         # healthd
@@ -90,12 +96,16 @@ class Deployer:
             'networking.istio.io', 'v1alpha3', 'default', 'virtualservices',
             specs.healthd_virtual_service_spec)
 
-        os.system("kubectl create ns kafka")
-        os.system("kubectl apply -f 'https://strimzi.io/install/latest?namespace=kafka' -n kafka")
-        os.system("kubectl apply -f "
-                  "https://strimzi.io/examples/latest/kafka/kafka-persistent-single.yaml -n kafka ")
+        if namespace.kafka:
+            os.system("kubectl create ns kafka")
+            os.system("kubectl apply -f 'https://strimzi.io/install/latest?namespace=kafka' -n kafka")
+            os.system("kubectl apply -f "
+                "https://strimzi.io/examples/latest/kafka/kafka-persistent-single.yaml -n kafka ")
 
-    def delete(self, _):
+    def delete(self, namespace):
+        print("The deploy module is used for dev deployments. As such, we are now "
+              "setting the kube context to docker-desktop.", flush=True)
+        subprocess.check_output("kubectl config use-context docker-desktop".split())
         self.delete_namespaced_custom_object(
             'networking.istio.io', 'v1alpha3', 'default', 'virtualservices',
             'healthd')
@@ -119,7 +129,8 @@ class Deployer:
         self.delete_namespaced_service_account('rexflow-etcd', 'rexflow')
         self.delete_namespace('rexflow')
 
-        os.system("kubectl delete -f "
-                  "https://strimzi.io/examples/latest/kafka/kafka-persistent-single.yaml -n kafka ")
-        os.system("kubectl delete -f 'https://strimzi.io/install/latest?namespace=kafka' -n kafka")
-        os.system("kubectl delete ns kafka")
+        if namespace.kafka:
+            os.system("kubectl delete -f "
+                    "https://strimzi.io/examples/latest/kafka/kafka-persistent-single.yaml -n kafka ")
+            os.system("kubectl delete -f 'https://strimzi.io/install/latest?namespace=kafka' -n kafka")
+            os.system("kubectl delete ns kafka")
